# docs/api/api.yaml
openapi: 3.0.0
info:
  version: '1'
  title: Social API
  description: >
    The Social API is a comprehensive RESTful interface developed and

    maintained by the Tranvelocity team. It facilitates communication with the

    Social platform, allowing users to access and manage various

    social-related functionalities.


    Authentication and Authorization:


    The API employs a robust authentication and authorization mechanism to

    ensure secure access to resources. To interact with the API, users must

    include the following headers in their requests:


    - `X-Social-Api-Key`: A unique API key assigned to each user for

    authentication purposes.


    - `X-Social-Timestamp`: A Unix timestamp indicating the time of the

    request.


    - `X-Social-Signature`: A cryptographic signature generated using the

    `SHA-512` algorithm based on the API key, timestamp, and secret key. These

    headers are required for all requests.


    - `X-Social-User-SSID`: The User login session ID for user

    authentication. If provided, the API will specify the user's role based on

    this login session.


    - `X-Social-Auth-Id`: The authentication ID providing additional user

    identification.


    The Social API will specify the user's role based on this login session.
    If these `X-Social-User-SSID` and `X-Social-Auth-Id`

    headers are not included, it indicates that the request is sent by

    non-registered users.


    Role-Based Access Control:


...
servers:
...
paths:
  /restricted-users:
    get:
      summary: Retrieve a list of restricted users
      description: >
        - This endpoint retrieves a paginated list of restricted users based on
        the provided search criteria.

        - It is intended for use in the admin panel only.

        - The `nickname` and `avatar` properties are sourced from the CRM API
        using the provided User ID.

        - The Social API does not manage these properties. To modify the
        nickname and avatar, users must update their profile on the member
        profile page.
      tags:
        - User Account Restriction-related endpoints
      operationId: get-restricted-users-list
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: user_id
          in: query
          required: false
          description: The User ID of the restricted user
          schema:
            type: integer
            example: 16614
        - name: remarks
          in: query
          required: false
          description: Remarks or comments associated with the restricted user
          schema:
            type: string
            example: test
        - name: offset
          in: query
          required: false
          description: The number of items to skip before starting to return results
          schema:
            type: integer
            example: 0
        - name: limit
          in: query
          required: false
          description: The maximum number of restricted user items to return
          schema:
            type: integer
            example: 30
      responses:
        '200':
          description: A list of restricted users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Unique identifier for the restricted user
                          example: 01j93f55zj72wrzm6ma7h1mm19
                        admin_uuid:
                          type: string
                          description: The UUID of the admin who applied the restriction
                          example: c20a5c21-cb68-4c36-b737-aed397f1bd66
                        remarks:
                          type: string
                          description: Remarks or comments associated with the restriction
                          example: test
                        user_id:
                          type: integer
                          description: The User ID of the restricted user
                          example: 16614
                        nickname:
                          type: string
                          description: Nickname of the restricted user (sourced from CRM)
                          example: HocTran (poster)
                        avatar:
                          type: string
                          description: Avatar URL of the restricted user (sourced from CRM)
                          example: >-
                            https://123456789.cloudfront.net/crm/staging/member/avatar/3317c26b-c484-4c62-b558-651f6fd51579-1722936935.png
                        updated_at:
                          type: string
                          description: >-
                            The timestamp of when the restriction was last
                            updated
                          example: '2024-10-01 16:19:16'
                        created_at:
                          type: string
                          description: The timestamp of when the restriction was created
                          example: '2024-10-01 16:19:16'
                  success:
                    type: boolean
                    description: Indicates whether the request was successful
                    example: true
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 200
                  pagination:
                    type: object
                    properties:
                      offset:
                        type: integer
                        example: 0
                      limit:
                        type: integer
                        example: 30
                      total:
                        type: integer
                        description: The total number of restricted users available
                        example: 2
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
    post:
      summary: Create a new restricted user
      description: >
        - This endpoint allows the creation of a new restricted user.

        - It is intended for use in the admin panel only.

        - The `nickname` and `avatar` properties are sourced from the CRM API
        using the provided User ID.

        - The Social API does not manage these properties. To modify the
        nickname and avatar, users must update their profile on the member
        profile page.
      tags:
        - User Account Restriction-related endpoints
      operationId: create-restricted-user
      requestBody:
        description: The parameters required to create a new restricted user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  description: The User ID of the restricted user (required)
                  example: 16614
                remarks:
                  type: string
                  description: Optional remarks or comments associated with the restriction
                  example: test
              required:
                - user_id
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
      responses:
        '201':
          description: The restricted user was successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Unique identifier for the restricted user
                        example: 01j93ftnyjs3c0z7pa6afvhs01
                      admin_uuid:
                        type: string
                        description: The UUID of the admin who created the restriction
                        example: c20a5c21-cb68-4c36-b737-aed397f1bd66
                      remarks:
                        type: string
                        description: Remarks or comments associated with the restriction
                        example: test
                      user_id:
                        type: integer
                        description: The User ID of the restricted user
                        example: 16614
                      nickname:
                        type: string
                        description: Nickname of the restricted user (sourced from CRM)
                        example: HocTran (poster)
                      avatar:
                        type: string
                        description: Avatar URL of the restricted user (sourced from CRM)
                        example: >-
                          https://123456789.cloudfront.net/crm/staging/member/avatar/3317c26b-c484-4c62-b558-651f6fd51579-1722936935.png
                      updated_at:
                        type: string
                        description: The timestamp of when the restriction was last updated
                        example: '2024-10-01 16:31:01'
                      created_at:
                        type: string
                        description: The timestamp of when the restriction was created
                        example: '2024-10-01 16:31:01'
                  success:
                    type: boolean
                    description: Indicates whether the request was successful
                    example: true
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 201
        '400':
          description: Validation error when creating the restricted user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates that the request was unsuccessful
                    example: false
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 400
                  errors:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Detailed error message
                        example: required|The user id field is required.
                      validation:
                        type: array
                        items:
                          type: object
                          properties:
                            attribute:
                              type: string
                              description: The field that caused the validation error
                              example: user_id
                            errors:
                              type: array
                              items:
                                type: object
                                properties:
                                  key:
                                    type: string
                                    description: The type of validation error
                                    example: Required
                                  message:
                                    type: string
                                    description: Detailed error message
                                    example: >-
                                      required|The user id field is
                                      required.
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '409':
          description: Conflict error when the user is already restricted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates that the request was unsuccessful
                    example: false
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 409
                  errors:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Conflict error message
                        example: このユーザーは既に制限されています。
                      code:
                        type: integer
                        description: Error code for conflict issues
                        example: 409001
  '/restricted-users/{id}':
    get:
      summary: Retrieve a specific restricted user by ID
      description: >
        - This endpoint retrieves the details of a specific restricted user by
        their ID.

        - It is intended for use in the admin panel only.

        - The `nickname` and `avatar` properties are sourced from
        the CRM API using the provided User ID.

        - The Social API does not manage these properties. To modify the nickname and avatar, users must
        update their profile on the member profile page.
      tags:
        - User Account Restriction-related endpoints
      operationId: get-restricted-user
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the restricted user to retrieve.
          schema:
            type: string
            example: 01j93ftnyjs3c0z7pa6afvhs01
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: role
          in: query
          required: true
          description: The role of the user (only "poster" role is allowed)
          schema:
            type: string
            example: poster
      responses:
        '200':
          description: The restricted user resource was successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Unique identifier for the restricted user
                        example: 01j93ftnyjs3c0z7pa6afvhs01
                      admin_uuid:
                        type: string
                        description: The UUID of the admin who created the restriction
                        example: c20a5c21-cb68-4c36-b737-aed397f1bd66
                      remarks:
                        type: string
                        description: Remarks or comments associated with the restriction
                        example: test
                      user_id:
                        type: integer
                        description: The User ID of the restricted user
                        example: 16614
                      nickname:
                        type: string
                        description: Nickname of the restricted user (sourced from CRM)
                        example: HocTran (poster)
                      avatar:
                        type: string
                        description: Avatar URL of the restricted user (sourced from CRM)
                        example: >-
                          https://123456789.cloudfront.net/crm/staging/member/avatar/3317c26b-c484-4c62-b558-651f6fd51579-1722936935.png
                      updated_at:
                        type: string
                        description: The timestamp of when the restriction was last updated
                        example: '2024-10-01 16:31:01'
                      created_at:
                        type: string
                        description: The timestamp of when the restriction was created
                        example: '2024-10-01 16:31:01'
                  success:
                    type: boolean
                    description: Indicates whether the request was successful
                    example: true
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 200
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '404':
          description: The specified restricted user was not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates that the request was unsuccessful
                    example: false
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 404
                  errors:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Error message indicating the user was not found
                        example: ID 01j93ftnyjs3c0z7pa6afvhs011 に該当する制限ユーザーが見つかりませんでした。
                      code:
                        type: integer
                        description: Error code indicating the resource was not found
                        example: 404001
    put:
      summary: Update a specific restricted user by ID
      description: >
        - This endpoint updates the details of a specific restricted user by
        their ID.

        - It is intended for use in the admin panel only.

        - The `nickname` and `avatar` properties are sourced from the CRM API
        using the provided User ID.

        - The Social API does not manage these properties. To modify the
        nickname and avatar, users must update their profile on the member
        profile page.
      tags:
        - User Account Restriction-related endpoints
      operationId: update-restricted-user
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the restricted user to update.
          schema:
            type: string
            example: 01j93ftnyjs3c0z7pa6afvhs01
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
      requestBody:
        description: >-
          The request body containing the parameters to update a restricted
          user.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                remarks:
                  type: string
                  description: Remarks or comments associated with the restriction.
                  example: This is a sample
      responses:
        '200':
          description: The restricted user resource was successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Unique identifier for the restricted user
                        example: 01j93ftnyjs3c0z7pa6afvhs01
                      admin_uuid:
                        type: string
                        description: The UUID of the admin who updated the restriction
                        example: c20a5c21-cb68-4c36-b737-aed397f1bd66
                      remarks:
                        type: string
                        description: Remarks or comments associated with the restriction
                        example: This is a sample
                      user_id:
                        type: integer
                        description: The User ID of the restricted user
                        example: 16614
                      nickname:
                        type: string
                        description: Nickname of the restricted user (sourced from CRM)
                        example: HocTran (poster)
                      avatar:
                        type: string
                        description: Avatar URL of the restricted user (sourced from CRM)
                        example: >-
                          https://123456789.cloudfront.net/crm/staging/member/avatar/3317c26b-c484-4c62-b558-651f6fd51579-1722936935.png
                      updated_at:
                        type: string
                        description: The timestamp of when the restriction was last updated
                        example: '2024-10-01 17:06:08'
                      created_at:
                        type: string
                        description: The timestamp of when the restriction was created
                        example: '2024-10-01 16:31:01'
                  success:
                    type: boolean
                    description: Indicates whether the request was successful
                    example: true
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 200
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '404':
          description: The specified restricted user was not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates that the request was unsuccessful
                    example: false
                  code:
                    type: integer
                    description: The HTTP status code of the response
                    example: 404
                  errors:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Error message indicating the user was not found
                        example: ID 01j93h6tgqgb64v7sj8wp3hwt4 に該当する制限ユーザーが見つかりませんでした。
                      code:
                        type: integer
                        description: Error code indicating the resource was not found
                        example: 404001
    delete:
      summary: Delete a specific restricted user by ID
      description: >
        - This endpoint deletes a specific restricted user by their ID.

        - It is intended for use in the admin panel only.
      tags:
        - User Account Restriction-related endpoints
      operationId: delete-restricted-user
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the restricted user to delete.
          schema:
            type: string
            example: 01j93ftnyjs3c0z7pa6afvhs01
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
      responses:
        '204':
          description: The restricted user has been deleted successfully (no content).
          content:
            application/json:
              schema:
                type: string
              examples:
                no_content:
                  value: ''
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '404':
          description: The specified restricted user was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates that the request was unsuccessful.
                    example: false
                  code:
                    type: integer
                    description: The HTTP status code of the response.
                    example: 404
                  errors:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Error message indicating the user was not found.
                        example: ID 01j93dad5ptjtbg0p1gd7v8n60 に該当する制限ユーザーが見つかりませんでした。
                      code:
                        type: integer
                        description: Error code indicating the resource was not found.
                        example: 404001
...
components:
  parameters:
...
    X-Social-Api-Key:
      name: X-Social-Api-Key
      in: header
      description: |
        An unique string provided by Tranvelocity when you create an account. You
        must include this API key in the specified HTTP request header for every
        API request to authenticate your requests.
      required: true
      schema:
        type: string
        example: FSFSDFSDFDSpHYv7kMeb0yVkn2343242L091nI5GLkKPfAt234324bmNEaR
    X-Social-Timestamp:
      name: X-Social-Timestamp
      in: header
      description: |
        An Unix epoch value representing the current time. You must include this
        timestamp in the specified HTTP request header for every API request to
        prevent replay attacks and ensure request freshness.
      required: true
      schema:
        type: integer
        example: 1595907741
    X-Social-Signature:
      name: X-Social-Signature
      in: header
      description: >
        A signature that is HMAC-encoded with SHA-512 digest. You must include
        this

        signature in the specified HTTP request header for every API request to

        ensure data integrity and authenticity.
      required: true
      schema:
        type: string
        example: 5d714ce2385d1cf96260806a50f3f7e68b928cc8
    X-Social-User-SSID:
      name: X-Social-User-SSID
      in: header
      description: |
        The session ID for User.
      required: true
      schema:
        type: string
        example: MBhdHsd4ckmKNNDaScHCJjuPY5pgGFQOhMSSqPis
    X-Social-Auth-Id:
      name: X-Social-Auth-Id
      in: header
      description: |
        The ID used for authentication.
      required: true
      schema:
        type: string
        example: ebbae0f297e50fa92f2eaaa340a1f9125cad6e2d
...
  responses:
    401AuthenticationFailedResponse:
      description: Unauthorized. HMAC authentication failed or invalid session ID.
      content:
        application/json:
          examples:
            HMACAuthenticationFailed:
              summary: HMAC authentication failed.
              value:
                success: false
                code: 401
                errors:
                  message: >-
                    HMAC authentication was failed. Please make sure your
                    credentials again.
                  code: 401001
            InvalidSessionID:
              summary: Invalid session ID.
              value:
                success: false
                code: 401
                errors:
                  message: Invalid session ID.
                  code: 401004
...